#84: made non-blocking while loop to emit socket messages
